+++ app/src/main/java/com/icecrown/onyxridgecm/activities/MainActivity.java
@@ -1,5 +1,6 @@
 //**************************************************************
 // Project: OnyxRidge Construction Management
+//
 // File: MainActivity.java
 //
 // Written by: Raymond O'Neill