+++ app/src/main/java/com/icecrown/onyxridgecm/fragments/SettingsAndHelpFragment.java
@@ -1,3 +1,16 @@
+//**************************************************************
+// Project: OnyxRidge Construction Management
+//
+// File: SettingsAndHelpFragment.java
+//
+// Written by: Raymond O'Neill
+//
+// Written on: 11/4/2020
+//
+// Purpose: Used for displaying help and (potentially) settings
+//          information to the user to assist in app navigation
+//          and experience
+//**************************************************************
 package com.icecrown.onyxridgecm.fragments;
 
 import androidx.fragment.app.Fragment;