+++ app/src/main/java/com/icecrown/onyxridgecm/interfaces/IActiveFragmentChanged.java
@@ -1,7 +1,0 @@
-package com.icecrown.onyxridgecm.interfaces;
-
-import androidx.fragment.app.Fragment;
-
-public interface IActiveFragmentChanged {
-    void onActiveFragmentChanged(Fragment fragment);
-}